feat(items-layout): display items in both grid and table
